fix main and window gui, added database for user priv.
