--- importDB.py
@@ -1,1 +1,0 @@
-import csv, sqlite3
