+++ main_backend.py
@@ -1,7 +1,5 @@
 import sqlite3
 from tkinter import *
-from admin import Admin
-#from student import student
 def connect():
     conn=sqlite3.connect("database.db")
     cur=conn.cursor()
@@ -13,30 +11,34 @@ def check(username,password):
     conn=sqlite3.connect('database.db')
     cur = conn.cursor()
     if   (cur.execute('SELECT * FROM users WHERE username =? AND password = ?',(username,password))):
-        if cur.fetchone():
-            window = Tk()
-            window.title('Admin_User')
-            window.geometry('700x450')
-            obj=Admin(window)
-            window.mainloop()
-        else:
-            messagebox.showinfo('error','INVALID CREDENTIALS for ADMIN LOGIN')
+        result=cur.fetchone()
+        return result
 
-# def checks(name,password):                       # for student login
-#     conn=sqlite3.connect('login.db')
+    conn.commit()
+    conn.close()
+        # if cur.fetchone():
+        #     window = Tk()
+        #     window.title('Admin_User')
+        #     window.geometry('700x450')
+        #     obj=Admin(window)
+        #     window.mainloop()
+        # else:
+        #     messagebox.showinfo('error','INVALID CREDENTIALS for ADMIN LOGIN')
+
+# def checks(username,password):                       # for student login
+#     conn=sqlite3.connect('database.db')
 #     cur = conn.cursor()
-#     if   (cur.execute('SELECT * FROM user WHERE name = ? AND password = ?', (name, password))):
+#     if   (cur.execute('SELECT * FROM users WHERE username = ? AND password = ? AND privilege==3', (username, password))):
 #         if cur.fetchone():
 #             window = Tk()
 #             window.title('Student_User')
-#             window.geometry('700x400')
+#             window.geometry('1400x500')
 #             obj = student(window)
 #             window.mainloop()
 #         else:
 #             messagebox.showinfo('error','INVALID CREDENTIALS for STUDENT LOGIN')
 #
-#
-#     conn.commit()
-#     conn.close()
+
+
 
 connect()
