+++ student.py
@@ -1,0 +1,115 @@
+from tkinter import *
+from tkinter import ttk, messagebox
+from tkinter.ttk import Combobox
+import backend
+
+class Student:
+    def __init__(self,window):
+        self.window = window
+        self.window.geometry('1400x600')
+        self.frame = Frame(self.window, width=700,height=400)
+        self.courseName = StringVar()
+        self.LecturerName = StringVar()
+        self.lbl1=Label(self.window, text='Course Name:')
+        self.lbl2=Label(self.window, text='Lecturer Name:')
+        self.courseName=Entry(self.window,textvariable=self.courseName)
+        self.LecturerName=Entry(self.window,textvariable=self.LecturerName)
+        self.btn1 = Button(self.window, text='Submit',command=self.submit)
+        self.lbl1.place(x=100, y=50)
+        self.courseName.place(x=200, y=50)
+        self.lbl2.place(x=380, y=50)
+        self.LecturerName.place(x=485, y=50)
+        self.data=("", "Exam", "Pre-Exam")
+        self.cb=Combobox(self.window, values=self.data)
+        self.cb.place(x=685, y=50)
+        self.data1=("","1-Easy",'2','3','4','5','6',"7-Very Hard")
+        self.cb2=Combobox(self.window,values=self.data1)
+        self.cb2.place(x=885,y=50)
+        self.data2=("", "Answer", "No-Answer")
+        self.cb2=Combobox(self.window, values=self.data2)
+        self.cb2.place(x=1085, y=50)
+        self.btn1.place(x=1285,y=50)
+        self.searchResults=Label(self.window,text="Search Results: ")
+        self.searchResults.place(x=100,y=150)
+        self.listbox=Listbox(self.window)
+        self.listbox.place(height=200,width=1185,x=100,y=200)
+
+        # self.label = Label(self.frame,text='Student User',font=('Georgia',30,'bold'))
+        # self.label.place(x=20,y=20,width=400,height=50)
+
+        # self.label_title = Label(self.frame, text='TITLE',font=('Georgia',14,'bold'))
+        # self.label_title.place(x=20,y=100,width=100,height=50)
+
+        # self.label_year = Label(self.frame, text='YEAR',font=('Georgia',14,'bold'))
+        # self.label_year.place(x=20,y=150,width=100,height=30)
+
+        # self.label_author = Label(self.frame, text='AUTHOR',font=('Georgia',14,'bold'))
+        # self.label_author.place(x=350,y=100,width=100,height=30)
+
+        # self.label_isbn = Label(self.frame, text='ISBN',font=('Georgia',14,'bold'))
+        # self.label_isbn.place(x=350,y=150,width=100,height=30)
+
+        # self.title_text=StringVar()
+        # self.entry_title = Entry(self.frame, fg='gray',textvariable=self.title_text,width=25,font=('Arial',12,'bold'))
+        # self.entry_title.place(x=120,y=100,width=150,height=30)
+
+        # self.year_text=StringVar()
+        # self.entry_year = Entry(self.frame, fg='gray',textvariable=self.year_text,width=25,font=('Arial',12,'bold'))
+        # self.entry_year.place(x=120,y=150,width=150,height=30)
+
+        # self.author_text=StringVar()
+        # self.entry_author = Entry(self.frame, fg='gray',textvariable=self.author_text,width=25,font=('Arial',12,'bold'))
+        # self.entry_author.place(x=470,y=100,width=150,height=30)
+
+        # self.isbn_text=StringVar()
+        # self.entry_isbn = Entry(self.frame, fg='gray',textvariable=self.isbn_text,width=25,font=('Arial',12,'bold'))
+        # self.entry_isbn.place(x=470,y=150,width=150,height=30)
+
+        # self.listbox = Listbox(self.frame)
+        # self.listbox.place(x=100,y=200,width=500,height=100)
+
+        # self.button_view = Button(self.frame,text='View All', command=self.view_command)
+        # self.button_view.place(x=100,y=320,width=100,height=40)
+
+        # self.button_search = Button(self.frame,text='Search ', command=self.search_command)
+        # self.button_search.place(x=200,y=320,width=100,height=40)
+
+        # self.button_issue = Button(self.frame,text='Issue', command=self.issue_command)
+        # self.button_issue.place(x=300,y=320,width=100,height=40)
+
+        # self.button_request = Button(self.frame, text='Request', command = self.request_command)
+        # self.button_request.place(x=400, y=320,width=100,height=40)
+
+        # self.button_issue = Button(self.frame, text='Clear Fields', command=self.clear_command)
+        # self.button_issue.place(x=500, y=320,width=100,height=40)
+
+        self.frame.pack()
+    def submit(self):
+        pass
+    # def clear_command(self):
+    #     self.entry_title.delete(0,END)
+    #     self.entry_year.delete(0,END)
+    #     self.entry_author.delete(0,END)
+    #     self.entry_isbn.delete(0,END)
+    #
+    # def view_command(self):
+    #     self.listbox.delete(0,END)
+    #     for row in backend.view():
+    #         self.listbox.insert(END,row)
+    #
+    # def search_command(self):
+    #     self.listbox.delete(0,END)
+    #     for row in backend.search(self.title_text.get(),self.author_text.get(),self.year_text.get(),self.isbn_text.get()):
+    #         self.listbox.insert(END,row)
+    #
+    # def add_command(self):
+    #     backend.insert(self.title_text.get(),self.author_text.get(),self.year_text.get(),self.isbn_text.get())
+    #     self.listbox.delete(0,END)
+    #     self.listbox.insert(END,(self.title_text.get(),self.author_text.get(),self.year_text.get(),self.isbn_text.get()))
+
+    '''
+window = Tk()
+window.title('Student_User')
+window.geometry('700x400')
+obj = student(window)
+window.mainloop()'''
